fix: Moved some classes from Entities to WebServices-Models and changed namespace for them
