+++ TownComparisons/TownComparisons.Domain/Abstract/IService.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using TownComparisons.Domain.Entities;
+using TownComparisons.Domain.WebServices.Models;
 
 namespace TownComparisons.Domain.Abstract
 {
