+++ TownComparisons/TownComparisons.Domain/Abstract/TownWebServiceBase.cs
@@ -6,6 +6,7 @@ using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 using TownComparisons.Domain.Entities;
+using TownComparisons.Domain.WebServices.Models;
 
 namespace TownComparisons.Domain.Abstract
 {
