--- TownComparisons/TownComparisons.Domain/Entities/Kpi.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TownComparisons.Domain.Entities
-{
-    //public class Kpi
-    //{
-    //    public string Auspices { get; set; }
-    //    public string Description { get; set; }
-    //    public bool Has_ou_data { get; set; }
-    //    public string Id { get; set; }
-    //    public int Is_divided_by_gender { get; set; }
-    //    public string Municipality_type { get; set; }
-    //    public string Operating_area { get; set; }
-    //    public DateTime? Ou_publication_date { get; set; }
-    //    public string Perspective { get; set; }
-    //    public DateTime? Prel_publication_date { get; set; }
-    //    public DateTime? Publication_date { get; set; }
-    //    public string Publ_period { get; set; }
-    //    public string Title { get; set; }
-    //}
-}
