+++ TownComparisons/TownComparisons.Domain/TownComparisons.Domain.csproj
@@ -67,15 +67,14 @@
     <Compile Include="DAL\TownComparisonsInitializer.cs" />
     <Compile Include="DAL\UnitOfWork.cs" />
     <Compile Include="Entities\GroupCategory.cs" />
-    <Compile Include="Entities\Kpi.cs" />
-    <Compile Include="Entities\KpiAnswer.cs" />
-    <Compile Include="Entities\KpiAnswers.cs" />
-    <Compile Include="Entities\KpiAnswerValues.cs" />
-    <Compile Include="Entities\KpiGroup.cs" />
-    <Compile Include="Entities\KpiGroups.cs" />
-    <Compile Include="Entities\KpiQuestion.cs" />
+    <Compile Include="WebServices\Models\KpiAnswer.cs" />
+    <Compile Include="WebServices\Models\KpiAnswers.cs" />
+    <Compile Include="WebServices\Models\KpiAnswerValues.cs" />
+    <Compile Include="WebServices\Models\KpiGroup.cs" />
+    <Compile Include="WebServices\Models\KpiGroups.cs" />
+    <Compile Include="WebServices\Models\KpiQuestion.cs" />
     <Compile Include="Entities\OrganisationalUnitInfo.cs" />
-    <Compile Include="Entities\OrganisationalUnits.cs" />
+    <Compile Include="WebServices\Models\OrganisationalUnits.cs" />
     <Compile Include="Entities\WebServiceQuery.cs" />
     <Compile Include="SessionHandler.cs" />
     <Compile Include="SettingsForFile.cs" />
@@ -83,7 +82,7 @@
     <Compile Include="Entities\Location.cs" />
     <Compile Include="Entities\Municipality.cs" />
     <Compile Include="Entities\Operator.cs" />
-    <Compile Include="Entities\OrganisationalUnit.cs" />
+    <Compile Include="WebServices\Models\OrganisationalUnit.cs" />
     <Compile Include="Entities\Property.cs" />
     <Compile Include="Helpers\CacheManager.cs" />
     <Compile Include="Helpers\ICache.cs" />
