+++ TownComparisons/TownComparisons.Domain/WebServices/Models/KpiAnswers.cs
@@ -5,14 +5,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TownComparisons.Domain.Entities
+namespace TownComparisons.Domain.WebServices.Models
 {
-    public class OrganisationalUnits
+    public class KpiAnswers
     {
         public int Count { get; set; }
 
         [JsonProperty("values")]
-        public List<OrganisationalUnit> Values { get; set; }
-
+        public List<KpiAnswer> Values { get; set; }
     }
 }
