+++ TownComparisons/TownComparisons.Domain/WebServices/Models/KpiGroup.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TownComparisons.Domain.Entities
+namespace TownComparisons.Domain.WebServices.Models
 {
     public class KpiGroup
     {
