+++ TownComparisons/TownComparisons.Domain/WebServices/Models/OrganisationalUnit.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TownComparisons.Domain.Entities
+namespace TownComparisons.Domain.WebServices.Models
 {
     public class OrganisationalUnit
     {
