+++ TownComparisons/TownComparisons.MVC/Views/Home/Compare.cshtml
@@ -1,4 +1,5 @@
 ﻿@using TownComparisons.Domain.Entities;
+@using TownComparisons.Domain.WebServices.Models
 @{
 
     ViewBag.Title = "Jämför";
