+++ TownComparisons/TownComparisons.MVC/Views/Home/Test.cshtml
@@ -1,4 +1,5 @@
 ﻿@using TownComparisons.Domain.Entities;
+@using TownComparisons.Domain.WebServices.Models
 
 
 <h3>Välj en kategori</h3>
